Fix casting issues

+++ dmalloc_t.c
@@ -181,8 +181,8 @@ static	void	free_slot(const int iter_c, pnt_info_t *slot_p,
   pnt_info_t	*this_p, *prev_p;
   
   if (verbose_b) {
-    (void)printf("%d: free'd %d bytes from slot %d (%#lx)\n",
-		 iter_c + 1, slot_p->pi_size, slot_p - pointer_grid,
+    (void)printf("%d: free'd %d bytes from slot %ld (%#lx)\n",
+		 iter_c + 1, slot_p->pi_size, (long)(slot_p - pointer_grid),
 		 (long)slot_p->pi_pnt);
   }
   
@@ -289,8 +289,8 @@ static	int	do_random(const int iter_n)
       free_c++;
       
       if (verbose_b) {
-	(void)printf("%d: free'd %d bytes from slot %d (%#lx)\n",
-		     iter_c + 1, pnt_p->pi_size, pnt_p - pointer_grid,
+	(void)printf("%d: free'd %d bytes from slot %ld (%#lx)\n",
+		     iter_c + 1, pnt_p->pi_size, (long)(pnt_p - pointer_grid),
 		     (long)pnt_p->pi_pnt);
       }
       
@@ -324,8 +324,8 @@ static	int	do_random(const int iter_n)
       pnt_p->pi_pnt = malloc(amount);
       
       if (verbose_b) {
-	(void)printf("%d: malloc %d of max %d into slot %d.  got %#lx\n",
-		     iter_c + 1, amount, max_avail, pnt_p - pointer_grid,
+	(void)printf("%d: malloc %d of max %d into slot %ld.  got %#lx\n",
+		     iter_c + 1, amount, max_avail, (long)(pnt_p - pointer_grid),
 		     (long)pnt_p->pi_pnt);
       }
       break;
@@ -336,8 +336,8 @@ static	int	do_random(const int iter_n)
       pnt_p->pi_pnt = calloc(amount, sizeof(char));
       
       if (verbose_b) {
-	(void)printf("%d: calloc %d of max %d into slot %d.  got %#lx\n",
-		     iter_c + 1, amount, max_avail, pnt_p - pointer_grid,
+	(void)printf("%d: calloc %d of max %d into slot %ld.  got %#lx\n",
+		     iter_c + 1, amount, max_avail, (long)(pnt_p - pointer_grid),
 		     (long)pnt_p->pi_pnt);
       }
       
@@ -376,9 +376,9 @@ static	int	do_random(const int iter_n)
       max_avail += pnt_p->pi_size;
       
       if (verbose_b) {
-	(void)printf("%d: realloc %d from %d of max %d slot %d.  got %#lx\n",
+	(void)printf("%d: realloc %d from %d of max %d slot %ld.  got %#lx\n",
 		     iter_c + 1, amount, pnt_p->pi_size, max_avail,
-		     pnt_p - pointer_grid, (long)pnt_p->pi_pnt);
+		     (long)(pnt_p - pointer_grid), (long)pnt_p->pi_pnt);
       }
       
       if (amount == 0) {
@@ -408,9 +408,9 @@ static	int	do_random(const int iter_n)
       max_avail += pnt_p->pi_size;
       
       if (verbose_b) {
-	(void)printf("%d: recalloc %d from %d of max %d slot %d.  got %#lx\n",
+	(void)printf("%d: recalloc %d from %d of max %d slot %ld.  got %#lx\n",
 		     iter_c + 1, amount, pnt_p->pi_size, max_avail,
-		     pnt_p - pointer_grid, (long)pnt_p->pi_pnt);
+		     (long)(pnt_p - pointer_grid), (long)pnt_p->pi_pnt);
       }
       
       /* test the returned block to make sure that is has been cleared */
@@ -442,8 +442,8 @@ static	int	do_random(const int iter_n)
       pnt_p->pi_pnt = valloc(amount);
       
       if (verbose_b) {
-	(void)printf("%d: valloc %d of max %d into slot %d.  got %#lx\n",
-		     iter_c + 1, amount, max_avail, pnt_p - pointer_grid,
+	(void)printf("%d: valloc %d of max %d into slot %ld.  got %#lx\n",
+		     iter_c + 1, amount, max_avail, (long)(pnt_p - pointer_grid),
 		     (long)pnt_p->pi_pnt);
       }
       break;
@@ -496,8 +496,8 @@ static	int	do_random(const int iter_n)
 	
 	if (verbose_b) {
 	  /* the amount includes the \0 */
-	  (void)printf("%d: strdup %d of max %d into slot %d.  got %#lx\n",
-		       iter_c + 1, amount + 1, max_avail, pnt_p - pointer_grid,
+	  (void)printf("%d: strdup %d of max %d into slot %ld.  got %#lx\n",
+		       iter_c + 1, amount + 1, max_avail, (long)(pnt_p - pointer_grid),
 		       (long)pnt_p->pi_pnt);
 	}
       }
@@ -516,8 +516,8 @@ static	int	do_random(const int iter_n)
 	pnt_p->pi_pnt = strndup(str, amount);
 	
 	if (verbose_b) {
-	  (void)printf("%d: strdup %d of max %d into slot %d.  got %#lx\n",
-		       iter_c + 1, amount, max_avail, pnt_p - pointer_grid,
+	  (void)printf("%d: strdup %d of max %d into slot %ld.  got %#lx\n",
+		       iter_c + 1, amount, max_avail, (long)(pnt_p - pointer_grid),
 		       (long)pnt_p->pi_pnt);
 	}
       }
@@ -3275,8 +3275,8 @@ static	int	check_special(void)
     
     else if (user_size != (SIZE_ARG) * (COUNT_ARG)) {
       if (! silent_b) {
-	(void)printf("   ERROR: calloc size should be %d but was %d.\n",
-		     (SIZE_ARG) * (COUNT_ARG), user_size);
+	(void)printf("   ERROR: calloc size should be %d but was %ld.\n",
+		     (SIZE_ARG) * (COUNT_ARG), (long)user_size);
       }
       final = 0;
     }
@@ -3924,43 +3924,43 @@ static	void	track_alloc_trxn(const char *file, const unsigned int line,
   
   switch (func_id) {
   case DMALLOC_FUNC_MALLOC:
-    (void)printf("%s malloc %d bytes got %#lx\n",
-		 file_line, byte_size, (long)new_addr);
+    (void)printf("%s malloc %ld bytes got %#lx\n",
+		 file_line, (long)byte_size, (long)new_addr);
     break;
   case DMALLOC_FUNC_CALLOC:
-    (void)printf("%s calloc %d bytes got %#lx\n",
-		 file_line, byte_size, (long)new_addr);
+    (void)printf("%s calloc %ld bytes got %#lx\n",
+		 file_line, (long)byte_size, (long)new_addr);
     break;
   case DMALLOC_FUNC_REALLOC:
-    (void)printf("%s realloc %d bytes from %#lx got %#lx\n",
-		 file_line, byte_size, (long)old_addr, (long)new_addr);
+    (void)printf("%s realloc %ld bytes from %#lx got %#lx\n",
+		 file_line, (long)byte_size, (long)old_addr, (long)new_addr);
     break;
   case DMALLOC_FUNC_RECALLOC:
-    (void)printf("%s recalloc %d bytes from %#lx got %#lx\n",
-		 file_line, byte_size, (long)old_addr, (long)new_addr);
+    (void)printf("%s recalloc %ld bytes from %#lx got %#lx\n",
+		 file_line, (long)byte_size, (long)old_addr, (long)new_addr);
     break;
   case DMALLOC_FUNC_MEMALIGN:
-    (void)printf("%s memalign %d bytes alignment %d got %#lx\n",
-		 file_line, byte_size, alignment, (long)new_addr);
+    (void)printf("%s memalign %ld bytes alignment %ld got %#lx\n",
+		 file_line, (long)byte_size, (long)alignment, (long)new_addr);
     break;
   case DMALLOC_FUNC_VALLOC:
-    (void)printf("%s valloc %d bytes alignment %d got %#lx\n",
-		 file_line, byte_size, alignment, (long)new_addr);
+    (void)printf("%s valloc %ld bytes alignment %ld got %#lx\n",
+		 file_line, (long)byte_size, (long)alignment, (long)new_addr);
     break;
   case DMALLOC_FUNC_STRDUP:
-    (void)printf("%s strdup %d bytes ot %#lx\n",
-		 file_line, byte_size, (long)new_addr);
+    (void)printf("%s strdup %ld bytes ot %#lx\n",
+		 file_line, (long)byte_size, (long)new_addr);
     break;
   case DMALLOC_FUNC_FREE:
     (void)printf("%s free %#lx\n", file_line, (long)old_addr);
     break;
   case DMALLOC_FUNC_NEW:
-    (void)printf("%s new %d bytes got %#lx\n",
-		 file_line, byte_size, (long)new_addr);
+    (void)printf("%s new %ld bytes got %#lx\n",
+		 file_line, (long)byte_size, (long)new_addr);
     break;
   case DMALLOC_FUNC_NEW_ARRAY:
-    (void)printf("%s new[] %d bytes got %#lx\n",
-		 file_line, byte_size, (long)new_addr);
+    (void)printf("%s new[] %ld bytes got %#lx\n",
+		 file_line, (long)byte_size, (long)new_addr);
     break;
   case DMALLOC_FUNC_DELETE:
     (void)printf("%s delete %#lx\n", file_line, (long)old_addr);
@@ -3969,9 +3969,9 @@ static	void	track_alloc_trxn(const char *file, const unsigned int line,
     (void)printf("%s delete[] %#lx\n", file_line, (long)old_addr);
     break;
   default:
-    (void)printf("%s unknown function %d bytes, %d alignment, %#lx old-addr "
+    (void)printf("%s unknown function %ld bytes, %ld alignment, %#lx old-addr "
 		 "%#lx new-addr\n",
-		 file_line, byte_size, alignment, (long)old_addr,
+		 file_line, (long)byte_size, (long)alignment, (long)old_addr,
 		 (long)new_addr);
     break;
   }
